requirements.txt
google.generativeai==0.2.2
weaviate-client==4.3b0
